overmount the container root
this fixes the weird corner case where customers running docker
inside a pvm would report that /cache/containers magically shows up in their mount paths.

the rabbit hole is here
https://elixir.bootlin.com/linux/latest/source/fs/namespace.c#L4983

setns changes the current task file system root to the namespaces root,
which is what we want.

however, the namespace root is NOT the root of the process that creates the namespace.
it's a separate tree, which is NOT changed by chroot or pivot_root.
it is changed only by mount.